cast int

+++ image_processing.py
@@ -118,5 +118,5 @@ def get_total_length(image_path: str, channel: str = 'k', kernel_size: int = 20)
     segment_list = get_segment_list(image_path, channel, kernel_size)
 
     # Get the sum of segment lengths
-    return sum(segment_list)
+    return (int)sum(segment_list)
     #return get_height_pix(image_path, 5, channel, kernel_size)
